requirements.txt
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
CacheControl==0.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.1
firebase==4.0.1
firebase_admin==7.1.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.22.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==26.0
pillow==11.3.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
psycopg[binary]
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
opencv-python-headless
numpy
